refactor: globalState 폴더명 store로 변경
